Refactoring build to run and buildTask to task
