+++ commands/azureCommands/acr-logs-utils/logScripts.js
