+++ commands/azureCommands/acr-logs-utils/resizable.js
