+++ commands/azureCommands/acr-logs-utils/style/fabric-components/css/vscmdl2-icons.css
