+++ commands/azureCommands/acr-logs-utils/style/fabric-components/fonts/vscmdl2-icons.ttf
