+++ commands/azureCommands/acr-logs-utils/style/stylesheet.css
