+++ commands/azureCommands/acr-logs-utils/tableDataManager.ts
