+++ commands/azureCommands/show-task.ts
@@ -1,43 +1,34 @@
-import { TaskRunRequest } from "azure-arm-containerregistry/lib/models";
 import { Registry } from "azure-arm-containerregistry/lib/models";
 import { ResourceGroup } from "azure-arm-resource/lib/resource/models";
 import { Subscription } from "azure-arm-resource/lib/subscription/models";
-import vscode = require('vscode');
 import { TaskNode } from "../../explorer/models/taskNode";
 import { ext } from '../../extensionVariables';
 import * as acrTools from '../../utils/Azure/acrTools';
 import { AzureUtilityManager } from "../../utils/azureUtilityManager";
 import { quickPickACRRegistry, quickPickSubscription, quickPickTask } from '../utils/quick-pick-azure';
+import { openTask } from "./task-utils/showTaskManager";
 
-export async function runBuildTask(context?: TaskNode): Promise<any> {
-    let taskName: string;
+export async function showTaskProperties(context?: TaskNode): Promise<any> {
     let subscription: Subscription;
-    let resourceGroup: ResourceGroup;
     let registry: Registry;
+    let resourceGroup: ResourceGroup;
+    let task: string;
 
-    if (context) { // Right Click
+    if (context) { // Right click
         subscription = context.subscription;
         registry = context.registry;
         resourceGroup = await acrTools.getResourceGroup(registry, subscription);
-        taskName = context.task.name;
-    } else { // Command Palette
+        task = context.task.name;
+    } else { // Command palette
         subscription = await quickPickSubscription();
         registry = await quickPickACRRegistry();
         resourceGroup = await acrTools.getResourceGroup(registry, subscription);
-        taskName = (await quickPickTask(registry, subscription, resourceGroup)).name;
+        task = (await quickPickTask(registry, subscription, resourceGroup)).name;
     }
 
     const client = AzureUtilityManager.getInstance().getContainerRegistryManagementClient(subscription);
-    let runRequest: TaskRunRequest = {
-        type: 'TaskRunRequest',
-        taskName: taskName
-    };
-
-    try {
-        let taskRun = await client.registries.scheduleRun(resourceGroup.name, registry.name, runRequest);
-        vscode.window.showInformationMessage(`Successfully ran the Task: ${taskName} with ID: ${taskRun.runId}`);
-    } catch (err) {
-        ext.outputChannel.append(err);
-        vscode.window.showErrorMessage(`Failed to ran the Task: ${taskName}`);
-    }
+    let item: any = await client.tasks.get(resourceGroup.name, registry.name, task);
+    let indentation = 1;
+    let replacer;
+    openTask(JSON.stringify(item, replacer, indentation), task);
 }
